manager workflow cleanup

+++ pygeoapi/process/manager/tinydb_.py
@@ -2,7 +2,7 @@
 #
 # Authors: Tom Kralidis <tomkralidis@gmail.com>
 #
-# Copyright (c) 2019 Tom Kralidis
+# Copyright (c) 2020 Tom Kralidis
 #
 # Permission is hereby granted, free of charge, to any person
 # obtaining a copy of this software and associated documentation
@@ -52,9 +52,7 @@ class TinyDBManager(BaseManager):
         :returns: `pygeoapi.process.manager.base.BaseManager`
         """
 
-        BaseManager.__init__(self, manager_def)
-
-        self.connection = manager_def['connection']
+        super().__init__(manager_def)
 
     def connect(self):
 
@@ -78,11 +76,11 @@ class TinyDBManager(BaseManager):
         self.db.close()
         return True
 
-    def get_jobs(self, processid=None, status=None):
+    def get_jobs(self, process_id=None, status=None):
         """
         Get jobs
 
-        :param processid: process identifier
+        :param process_id: process identifier
         :param status: job status (accepted, running, successful,
                        failed, results) (default is all)
 
@@ -90,13 +88,15 @@ class TinyDBManager(BaseManager):
         """
 
         self.connect()
-        if processid is None:
-            return [doc.doc_id for doc in self.db.all()]
+        if process_id is None:
+            jobs_list = [doc.doc_id for doc in self.db.all()]
         else:
             query = tinydb.Query()
-            return self.db.search(query.processid == processid)
+            jobs_list = self.db.search(query.process_id == process_id)
 
         self.db.close()
+
+        return jobs_list
 
     def add_job(self, job_metadata):
         """
@@ -104,19 +104,20 @@ class TinyDBManager(BaseManager):
 
         :param job_metadata: `dict` of job metadata
 
-        :returns: `bool` of add job result
+        :returns: identifier of added job
         """
 
         self.connect()
         doc_id = self.db.insert(job_metadata)
         self.db.close()
+
         return doc_id
 
-    def update_job(self, processid, job_id, update_dict):
+    def update_job(self, process_id, job_id, update_dict):
         """
         Updates a job
 
-        :param processid: process identifier
+        :param process_id: process identifier
         :param job_id: job identifier
         :param update_dict: `dict` of property updates
 
@@ -126,22 +127,23 @@ class TinyDBManager(BaseManager):
         self.connect()
         self.db.update(update_dict, tinydb.where('identifier') == job_id)
         self.db.close()
+
         return True
 
-    def delete_job(self, processid, job_id):
+    def delete_job(self, process_id, job_id):
         """
         Deletes a job
 
-        :param processid: process identifier
+        :param process_id: process identifier
         :param job_id: job identifier
 
         :return `bool` of status result
         """
         # delete result file if present
-        job_result = self.get_job_result(processid, job_id)
+        job_result = self.get_job_result(process_id, job_id)
         if job_result:
             location = job_result.get('location', None)
-            if location:
+            if location and self.output_dir is not None:
                 os.remove(location)
 
         self.connect()
@@ -150,11 +152,11 @@ class TinyDBManager(BaseManager):
 
         return removed
 
-    def get_job_result(self, processid, job_id):
+    def get_job_result(self, process_id, job_id):
         """
         Get a single job
 
-        :param processid: process identifier
+        :param process_id: process identifier
         :param jobid: job identifier
 
         :returns: `dict`  # `pygeoapi.process.manager.Job`
@@ -162,21 +164,24 @@ class TinyDBManager(BaseManager):
 
         self.connect()
         query = tinydb.Query()
-        result = self.db.search((query.processid == processid) & (query.identifier == job_id))
+        result = self.db.search((
+            query.process_id == process_id) & (query.identifier == job_id))
+
         result = result[0] if result else None
         self.db.close()
         return result
 
-    def get_job_output(self, processid, job_id):
+    def get_job_output(self, process_id, job_id):
         """
         Get a job's status, and actual output of executing the process
 
-        :param processid: process identifier
+        :param process_id: process identifier
         :param jobid: job identifier
 
         :returns: tuple of JobStatus and the process output as a `dict`
         """
-        job_result = self.get_job_result(processid, job_id)
+
+        job_result = self.get_job_result(process_id, job_id)
         if not job_result:
             # processs/job does not exist
             return None, None
@@ -189,8 +194,9 @@ class TinyDBManager(BaseManager):
             # Job data was not written for some reason
             # TODO log/raise exception?
             return job_status, {}
-        with io.open(location, 'r') as filehandler:
+        with io.open(location, 'r', encoding='utf-8') as filehandler:
             result = json.load(filehandler)
+
         return job_status, result
 
     def __repr__(self):
